Extract table w/logic to separate component. Add "new" label to description column.
